--- a/src/datascript/db.cljc
+++ b/src/datascript/db.cljc
@@ -1,4 +1,5 @@
 (ns ^:no-doc ^:lean-ns datascript.db
+  #?(:clj (:refer-clojure :exclude [update]))
   (:require
     #?(:cljs [goog.array :as garray])
     [clojure.walk]
@@ -8,6 +9,9 @@
     [me.tonsky.persistent-sorted-set.arrays :as arrays])
   #?(:cljs (:require-macros [datascript.db :refer [case-tree combine-cmp raise defrecord-updatable cond+]]))
   (:refer-clojure :exclude [seqable?])) 
+
+#?(:clj (ns-unmap *ns* 'update))
+#?(:clj (require '[datascript.inline :refer [update]]))
 
 #?(:clj (set! *warn-on-reflection* true))