@tonsky 2022-08-24
Thanks! Merged and released as 1.3.15.

My results (Apple M1 Pro, Azul JVM 18):

<img width="517" alt="Screenshot 2022-08-24 at 21 14 52" src="https://user-images.githubusercontent.com/285292/186504020-098d7a92-0a31-4388-8dd4-82989f69cd04.png">

JVM only, CLJS is more or less the same